#92 Move Invite-Participants button to its own component
