+++ quiz-master-client/src/App.scss
@@ -1,5 +1,8 @@
+@import "./index.scss";
+
+
 .App {
-  min-height: 100vh;
+  min-height: calc(100vh - $navbar-height);
   display: flex;
   flex-direction: column;
   justify-content: flex-start;
@@ -34,7 +37,13 @@
 
 }
 
-.Quiz-dashboard .question, .Quiz-dashboard .participants {
+.Quiz-dashboard .question { 
+  padding: 15px;
+  padding-top: 9px;
+  margin: 10px;
+}
+
+.Quiz-dashboard .participants {
   padding: 15px;
   margin: 10px;
 }
