--- quiz-master-client/src/quiz-client-shared/AppHeader/AppHeader.css
@@ -1,17 +1,0 @@
-
-.navbar-brand {
-    font-size: 20px;
-    font-weight: 700;
-  }
-  
-  .app-logo {
-    width: 50px;
-    font-size: 30px;
-    padding-top: 30px;
-    margin-left: 10px;
-
-  }
-
-  .app-title {
-    padding-top: 15px;
-  }
