+++ quiz-master-client/src/quiz-client-shared/AppHeader/AppHeader.scss
@@ -1,0 +1,17 @@
+
+.navbar-brand {
+    .app-logo {
+      width: 50px;
+      font-size: 30px;
+      padding-top: 30px;
+      margin-left: 10px;
+    }
+  
+    .app-title {
+      padding-top: 15px;
+      font-size: 20px;
+      font-weight: 700;
+    }
+  }
+  
+
